ch: refactor out hand
 - refactor out hand

[Finishes #172825140]
